Renames MarkovModel to Markovnetwork
