+++ pgmpy/estimators/base.py
@@ -185,7 +185,7 @@ class ParameterEstimator(BaseEstimator):
 
         Parameters
         ----------
-        model: pgmpy.models.BayesianNetwork or pgmpy.models.MarkovModel or pgmpy.models.NoisyOrModel model
+        model: pgmpy.models.BayesianNetwork or pgmpy.models.MarkovNetwork or pgmpy.models.NoisyOrModel model
             for which parameter estimation is to be done.
 
         data: pandas DataFrame object
