+++ pgmpy/inference/mplp.py
@@ -5,7 +5,7 @@ import numpy as np
 import networkx as nx
 
 from pgmpy.inference import Inference
-from pgmpy.models import MarkovModel
+from pgmpy.models import MarkovNetwork
 from pgmpy.factors.discrete import DiscreteFactor
 
 
@@ -18,15 +18,15 @@ class Mplp(Inference):
 
     Parameters
     ----------
-    model: MarkovModel for which inference is to be performed.
+    model: MarkovNetwork for which inference is to be performed.
 
     Examples
     --------
     >>> import numpy as np
-    >>> from pgmpy.models import MarkovModel
+    >>> from pgmpy.models import MarkovNetwork
     >>> from pgmpy.inference import Mplp
     >>> from pgmpy.factors.discrete import DiscreteFactor
-    >>> student = MarkovModel()
+    >>> student = MarkovNetwork()
     >>> student.add_edges_from([('A', 'B'), ('B', 'C'), ('C', 'D'), ('E', 'F')])
     >>> factor_a = DiscreteFactor(['A'], cardinality=[2], values=np.array([0.54577, 1.8323]))
     >>> factor_b = DiscreteFactor(['B'], cardinality=[2], values=np.array([0.93894, 1.065]))
@@ -48,8 +48,8 @@ class Mplp(Inference):
     """
 
     def __init__(self, model):
-        if not isinstance(model, MarkovModel):
-            raise TypeError("Only MarkovModel is supported")
+        if not isinstance(model, MarkovNetwork):
+            raise TypeError("Only MarkovNetwork is supported")
 
         super(Mplp, self).__init__(model)
         self._initialize_structures()
@@ -316,10 +316,10 @@ class Mplp(Inference):
 
         Examples
         --------
-        >>> from pgmpy.models import MarkovModel
+        >>> from pgmpy.models import MarkovNetwork
         >>> from pgmpy.factors.discrete import DiscreteFactor
         >>> from pgmpy.inference import Mplp
-        >>> mm = MarkovModel()
+        >>> mm = MarkovNetwork()
         >>> mm.add_nodes_from(['x1', 'x2', 'x3', 'x4', 'x5', 'x6', 'x7'])
         >>> mm.add_edges_from([('x1', 'x3'), ('x1', 'x4'), ('x2', 'x4'),
         ...                    ('x2', 'x5'), ('x3', 'x6'), ('x4', 'x6'),
@@ -477,10 +477,10 @@ class Mplp(Inference):
 
         Examples
         --------
-        >>> from pgmpy.models import MarkovModel
+        >>> from pgmpy.models import MarkovNetwork
         >>> from pgmpy.factors.discrete import DiscreteFactor
         >>> from pgmpy.inference import Mplp
-        >>> mm = MarkovModel()
+        >>> mm = MarkovNetwork()
         >>> mm.add_nodes_from(['x1', 'x2', 'x3', 'x4', 'x5', 'x6', 'x7'])
         >>> mm.add_edges_from([('x1', 'x3'), ('x1', 'x4'), ('x2', 'x4'),
         ...                    ('x2', 'x5'), ('x3', 'x6'), ('x4', 'x6'),
@@ -548,11 +548,11 @@ class Mplp(Inference):
 
         Examples
         --------
-        >>> from pgmpy.models import MarkovModel
+        >>> from pgmpy.models import MarkovNetwork
         >>> from pgmpy.factors.discrete import DiscreteFactor
         >>> from pgmpy.inference import Mplp
         >>> import numpy as np
-        >>> student = MarkovModel()
+        >>> student = MarkovNetwork()
         >>> student.add_edges_from([('A', 'B'), ('B', 'C'), ('C', 'D'), ('E', 'F')])
         >>> factor_a = DiscreteFactor(['A'], cardinality=[2], values=np.array([0.54577, 1.8323]))
         >>> factor_b = DiscreteFactor(['B'], cardinality=[2], values=np.array([0.93894, 1.065]))
