+++ pgmpy/models/__init__.py
@@ -5,6 +5,7 @@ from .DynamicBayesianNetwork import DynamicBayesianNetwork
 from .FactorGraph import FactorGraph
 from .JunctionTree import JunctionTree
 from .MarkovChain import MarkovChain
+from .MarkovNetwork import MarkovNetwork
 from .MarkovModel import MarkovModel
 from .NaiveBayes import NaiveBayes
 from .NoisyOrModel import NoisyOrModel
@@ -15,6 +16,7 @@ __all__ = [
     "BayesianModel",
     "BayesianNetwork",
     "NoisyOrModel",
+    "MarkovNetwork",
     "MarkovModel",
     "FactorGraph",
     "JunctionTree",
